QualityModelDeriver.deriveModel() tests written up to assertions
